完成增量更新接口对接oa并完成基本测试，个别bean字段增加值注释，注释规范等

+++ src/main/java/com/oce/base/controller/UserController.java
@@ -25,7 +25,7 @@ public class UserController {
         return userService.findAll();
     }
     @RequestMapping(value = "/user/queryByid/{id}",method = RequestMethod.GET)
-    public TUser queryByid(@PathVariable("id") int id){
+    public TUser queryByid(@PathVariable("id") long id){
         return userService.selectUserByid(id);
     }
     @RequestMapping(value = "/user/queryByname/{lastName}",method = RequestMethod.GET)
